move MapSocketError to a common file and fix some review issues
